requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pydantic
pillow
pytesseract
pdfplumber
jinja2